% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erlang_c_app.R
\name{erlang_c_app}
\alias{erlang_c_app}
\title{Shiny App for Erlang C Calculator}
\usage{
erlang_c_app(language = "en")
}
\arguments{
\item{language}{Character. The language to use for translations (default: "en").}
}
\value{
A Shiny app object.
}
\description{
This function creates a Shiny app for calculating Erlang C metrics.
}
