% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{translations}
\alias{translations}
\title{Translations}
\format{
A data frame with variables:
\describe{
  \item{key}{The key of the string to translate}
  \item{en}{The translation in English}
  \item{zh}{The translation in Chinese}
}
}
\source{
Local
}
\usage{
translations
}
\description{
This dataset contains translations for all the
strings used in the app. It is used to create
a Shiny string translation interface.
}
\keyword{datasets}
